% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqar.R
\name{iqar}
\alias{iqar}
\title{Compute IQAr index considering several air pollutants}
\usage{
iqar(
  pm10 = NA,
  pm2.5 = NA,
  o3 = NA,
  co = NA,
  no2 = NA,
  so2 = NA,
  label = FALSE
)
}
\arguments{
\item{pm10}{numeric. Particulate matter \eqn{PM_{10}} concentration, \eqn{\mu g/m^{3}}.}

\item{pm2.5}{numeric. Particulate matter \eqn{PM_{2.5}} concentration, \eqn{\mu g/m^{3}}.}

\item{o3}{numeric. Ozone \eqn{O_3} concentration, \eqn{\mu g/m^{3}}.}

\item{co}{numeric. Carbon monoxide \eqn{CO} concentration, \eqn{PPM}.}

\item{no2}{numeric. Nitrogen dioxide \eqn{NO_2} concentration, \eqn{\mu g/m^{3}}.}

\item{so2}{numeric. Sulfur dioxide \eqn{SO_2} concentration, \eqn{\mu g/m^{3}}.}

\item{label}{logical. If \code{FALSE}, the function will return the IQAr value, if \code{TRUE} the function will return the IQAr label. Defaults to \code{FALSE}.}
}
\value{
numeric. IQAr index value or classification.
}
\description{
This function computes the IQAr index considering several air pollutants using the Brazilian CONAMA/MMA specification.
}
\examples{
iqar(pm10 = 50, pm2.5 = 80, o3 = 60)
iqar(pm10 = 50, pm2.5 = 80, o3 = 60, label = TRUE)
}
