% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqar.R
\name{iqar}
\alias{iqar}
\title{Compute IQAr index considering several air pollutants}
\usage{
iqar(
  pm10 = NA,
  pm2.5 = NA,
  o3 = NA,
  co = NA,
  no2 = NA,
  so2 = NA,
  label = FALSE
)
}
\arguments{
\item{pm10}{numeric. Pollutant concentration. See details bellow for the correct unit.}

\item{pm2.5}{numeric. Pollutant concentration. See details bellow for the correct unit.}

\item{o3}{numeric. Pollutant concentration. See details bellow for the correct unit.}

\item{co}{numeric. Pollutant concentration. See details bellow for the correct unit.}

\item{no2}{numeric. Pollutant concentration. See details bellow for the correct unit.}

\item{so2}{numeric. Pollutant concentration. See details bellow for the correct unit.}

\item{label}{logical. If \code{FALSE}, the function will return the IQAr value, if \code{TRUE} the function will return the IQAr label. Defaults to \code{FALSE}.}
}
\value{
numeric. IQAr index value or classification.
}
\description{
This function computes the IQAr index considering several air pollutants using the Brazilian CONAMA/MMA specification.
}
\details{
Units: mp10: mug/m3, mp2.5: mug/m3, o3: mug/m3, co: PPM, no2: mug/m3, so2: mug/m3
}
\examples{
iqar(pm10 = 50, pm2.5 = 80, o3 = 60)
}
