% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqar_pol.R
\name{iqar_pol}
\alias{iqar_pol}
\title{Compute specific IQAr index to several air pollutants}
\usage{
iqar_pol(x, pol = "pm2.5", label = FALSE)
}
\arguments{
\item{x}{numeric. Pollutants concentration. See details bellow for the correct unit.}

\item{pol}{character. Pollutant abbreviation: pm10, pm2.5, o3, co, no2, so2.}

\item{label}{logical. If \code{FALSE}, the function will return the IQAr value, if \code{TRUE} the function will return the IQAr label. Defaults to \code{FALSE}.}
}
\value{
numeric. Pollutant specific IQAr index
}
\description{
This function computes specific IQAr index to several air pollutants using the Brazilian CONAMA/MMA specification.
}
\details{
Units: \eqn{PM_{10}}: \eqn{\mu g/m^{3}}, \eqn{PM_{2.5}}: \eqn{\mu g/m^{3}}, \eqn{O_3}: \eqn{\mu g/m^{3}}, \eqn{CO}: \eqn{PPM}, \eqn{NO_2}: \eqn{\mu g/m^{3}}, \eqn{SO_2}: \eqn{\mu g/m^{3}}
}
\examples{
iqar_pol(x = 10, pol = "pm2.5")
iqar_pol(x = 10, pol = "pm2.5", label = TRUE)
}
